refactor: lib/ を解消し savedQueries/settings/theme を domain/ に移動
lib/ に残っていた 3 つのファイル (savedQueries, settings, theme) を
domain/ 配下に移動。それぞれ chrome.storage を直接叩く部分が含まれるため
厳密には infra 寄りだが、port/adapter に分割するほどの実装の厚みがないため
複雑化を避けて移動のみとする。

- src/lib/savedQueries.ts → src/domain/savedQueries.ts
- src/lib/savedQueries.spec.ts → src/domain/savedQueries.spec.ts
- src/lib/settings.ts → src/domain/settings.ts
- src/lib/theme.ts → src/domain/theme.ts
- src/lib/__mocks__/chrome_storage.mock.ts → src/domain/__mocks__/

src/types/Settings.ts は別レイヤの共有型として現状維持。
