+++ src/components/settings/OptionsPage/index.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import { fn } from "storybook/test";
 
 import { OptionsPage } from "./index";
-import { DEFAULT_SETTINGS } from "../../../lib/settings";
+import { DEFAULT_SETTINGS } from "../../../domain/settings";
 
 const meta: Meta<typeof OptionsPage> = {
   component: OptionsPage,
