+++ src/options-main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 
 import "./index.css";
 import OptionsApp from "./OptionsApp";
-import { initializeTheme } from "./lib/theme";
+import { initializeTheme } from "./domain/theme";
 
 initializeTheme();
